unsafe WriteEncodedString again!
Updated to the latest master, and this time with a modified version of
your speed test that shows off the speed improvement rather than the
other way around! i.e. this is a performance improvement for bulk data,
rather that small calls. I believe this is more indicitive of the spirit
of the library (i.e. why would be bother optimizing functions for small
amounts of data??)