Allow field layout to be determined for more types.
Use the FormatterServices.GetUninitializedObject() method to get an
instance of a specified type without having to invoke one of the ctors
for the type; this allows the field layout to be determined for a wider
set of types, e.g., types whose ctors check-then-throw for null
arguments. An additional benefit of this change is that because the
ctors for those types are no longer run, they no longer raise
exceptions which cause the VS debugger to break (pause execution).