SDK .csproj, Multi-targeting, PackageReference, xunit
This does a lot of things. The root cause is issue #279, because Jil.dll and Jil.Core.dll differ anything restoring from a netstandard library that depends on Jil being restored to a net4x project go boom. Because the netstandard lib referenced Jil.Core, with a different signature. It cannot be fixed on the user end.

But hey, multi-targeting works now...so I upgraded Jil to use it. This commit:
- Drops JilCore and JilCore.Tests (folded into Jil and JilTests)
- Drops \packages (uses <PackageReference>)
- Adds a netstandard2.0 build (far fewer references and less baggage for most users)
- Moves testing to xunit (makes dotnet xunit available)
- Tests all 3 releases
- Adds a build.cmd (runs tests, creates packages in /.nupkgs/)
- Does *not* modify versioning, still manual.

To build, just run build.cmd.